Respawning Enemies and Float Health
Enemies can be set to respawn. After they die, 3 seconds later they will
respawn where they first spawned.
If there is no train engine set for the enemies, they just stand still.
This is going to be temporary for using them as target dummies.

health is now a float instead of an int. May go back and change it back to
an int at some point. I just made the damage for the guns be a float so
health is now a float too.

Made the enemy into a prefab and coloure it red.
Made the bullet projectile larger.

+++ Assets/Scripts/Health.cs
@@ -9,9 +9,9 @@ public class Health : MonoBehaviour {
     // player, enemies, train cars
 
 
-    public int baseHp = 100;
+    public float baseHp = 100;
     [SerializeField]
-    int currentHp;
+    float currentHp;
 
     public event Action OnDeath;
     public event Action OnDamage;
@@ -33,7 +33,7 @@ public class Health : MonoBehaviour {
         currentHp = baseHp;
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         if(OnDamage != null)
         {
